refactor and HistPlot class
